Update slotcheckerservice for global slot checker
The slotcheckerservice needs updating to talk to the new global
slot checker. Specifically:
    - the slot complete messages should indicate if this message
      originates from a component or a proxy
    - the component ID in the slot complete message should be a
      32-bit unsigned integer, rather than a 8-bit signed integer

Note that the IP that the slotcheckerservice connects to is hard
coded in the source (192.168.250.1). This was previously the IP
of the proxy, but with the split proxy design this needs to be the
IP address of the global slot checker.